+++ checkov/terraform/graph_builder/local_graph.py
@@ -28,6 +28,7 @@ from checkov.terraform.graph_builder.utils import (
 from checkov.terraform.graph_builder.utils import is_local_path
 from checkov.terraform.graph_builder.variable_rendering.renderer import TerraformVariableRenderer
 
+
 MODULE_RESERVED_ATTRIBUTES = ("source", "version")
 CROSS_VARIABLE_EDGE_PREFIX = '[cross-variable] '
 
@@ -62,9 +63,13 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             renderer = TerraformVariableRenderer(self)
             renderer.render_variables_from_local_graph()
             self.update_vertices_breadcrumbs_and_module_connections()
+            self.update_nested_modules_address()
             if strtobool(os.getenv("CHECKOV_EXPERIMENTAL_CROSS_VARIABLE_EDGES", "False")):
                 # experimental flag on building cross variable edges for terraform graph
+                logging.info("Building cross variable edges")
+                edges_count = len(self.edges)
                 self._build_cross_variable_edges()
+                logging.info(f"Found {len(self.edges) - edges_count} cross variable edges")
 
     def _create_vertices(self) -> None:
         logging.info("Creating vertices")
@@ -179,7 +184,9 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         for origin_node_index, vertex in enumerate(self.vertices):
             self._build_edges_for_vertex(origin_node_index, vertex, aliases, resources_types)
 
-    def _build_edges_for_vertex(self, origin_node_index: int, vertex: TerraformBlock, aliases: Dict[str, Dict[str, BlockType]], resources_types: List[str], cross_variable_edges: bool = False):
+    def _build_edges_for_vertex(self, origin_node_index: int, vertex: TerraformBlock, aliases: Dict[str, Dict[str, BlockType]], resources_types: List[str], cross_variable_edges: bool = False, referenced_module: Optional[Dict[str, Any]] = None):
+        referenced_module_idx = referenced_module.get("idx") if referenced_module else None
+        referenced_module_path = referenced_module.get("path") if referenced_module else None
         for attribute_key, attribute_value in vertex.attributes.items():
             if attribute_key in reserved_attribute_names or attribute_has_nested_attributes(
                     attribute_key, vertex.attributes
@@ -196,7 +203,12 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
                 sub_values = [remove_index_pattern_from_str(sub_value) for sub_value in vertex_reference.sub_parts]
                 for i in range(len(sub_values)):
                     reference_name = join_trimmed_strings(char_to_join=".", str_lst=sub_values, num_to_trim=i)
-                    if vertex.module_dependency:
+                    if referenced_module is not None:
+                        dest_node_index = self._find_vertex_index_relative_to_path(
+                            vertex_reference.block_type, reference_name, referenced_module_path, vertex.module_dependency,
+                            vertex.module_dependency_num, referenced_module_idx
+                        )
+                    elif vertex.module_dependency:
                         dest_node_index = self._find_vertex_index_relative_to_path(
                             vertex_reference.block_type, reference_name, vertex.path, vertex.module_dependency,
                             vertex.module_dependency_num
@@ -268,6 +280,9 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         for origin_node_index in target_nodes_indexes:
             vertex = self.vertices[origin_node_index]
             self._build_edges_for_vertex(origin_node_index, vertex, aliases, resources_types, True)
+            modules = vertex.breadcrumbs.get(CustomAttributes.SOURCE_MODULE, [])
+            for module in modules:
+                self._build_edges_for_vertex(origin_node_index, vertex, aliases, resources_types, True, module)
 
     def _create_edge(self, origin_vertex_index: int, dest_vertex_index: int, label: str,
                      cross_variable_edges: bool = False) -> bool:
@@ -275,9 +290,12 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             return False
         edge = Edge(origin_vertex_index, dest_vertex_index, label)
         if cross_variable_edges:
-            if edge in self.edges:
+            if edge in self.edges or self.vertices[edge.dest].block_type != BlockType.RESOURCE or \
+                    self.vertices[edge.origin].block_type != BlockType.RESOURCE:
                 return False
             edge.label = CROSS_VARIABLE_EDGE_PREFIX + edge.label
+            if edge in self.edges:
+                return False
         self.edges.append(edge)
         self.out_edges[origin_vertex_index].append(edge)
         self.in_edges[dest_vertex_index].append(edge)
@@ -350,10 +368,14 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
         return os.path.realpath(dest_module_path)
 
     def _find_vertex_index_relative_to_path(
-        self, block_type: BlockType, name: str, block_path: str, module_path: str, module_num: str
+        self, block_type: BlockType, name: str, block_path: str, module_path: str, module_num: str, relative_module_idx: Optional[int] = None
     ) -> int:
         relative_vertices = []
-        possible_vertices = self.vertices_by_module_dependency_by_name.get((module_path, module_num), {}).get(block_type, {}).get(name, [])
+        if relative_module_idx is not None:
+            module_dependency_by_name_key = next(k for k, v in self.vertices_by_module_dependency.items() if v.get(BlockType.MODULE, []).__contains__(relative_module_idx))
+        else:
+            module_dependency_by_name_key = (module_path, module_num)
+        possible_vertices = self.vertices_by_module_dependency_by_name.get(module_dependency_by_name_key, {}).get(block_type, {}).get(name, [])
         for vertex_index in possible_vertices:
             vertex = self.vertices[vertex_index]
             if self.get_dirname(vertex.path) == self.get_dirname(block_path):
@@ -464,11 +486,14 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
                     hash_breadcrumbs.append(breadcrumb)
                 vertex.breadcrumbs[attribute_key] = hash_breadcrumbs
             if len(vertex.source_module) == 1:
-                m = self.vertices[list(vertex.source_module)[0]]
-                source_module_data = [m.get_export_data()]
-                while len(m.source_module) == 1:
-                    m = self.vertices[list(m.source_module)[0]]
-                    source_module_data.append(m.get_export_data())
+                v = vertex
+                source_module_data = []
+                while len(v.source_module) == 1:
+                    idx = list(v.source_module)[0]
+                    v = self.vertices[idx]
+                    module_data = v.get_export_data()
+                    module_data["idx"] = idx
+                    source_module_data.append(module_data)
                 source_module_data.reverse()
                 vertex.breadcrumbs[CustomAttributes.SOURCE_MODULE] = source_module_data
 
@@ -499,6 +524,29 @@ class TerraformLocalGraph(LocalGraph[TerraformBlock]):
             dir_name = os.path.abspath(path)
             self.abspath_cache[path] = dir_name
         return dir_name
+
+    @staticmethod
+    def get_current_address(vertex: TerraformBlock, address_prefix: str = ''):
+        if vertex.block_type == BlockType.MODULE:
+            return address_prefix + f"{vertex.block_type}.{vertex.name}"
+        else:
+            return address_prefix + vertex.name
+
+    def update_nested_modules_address(self) -> None:
+        for vertex in self.vertices:
+            if vertex.block_type not in [BlockType.MODULE, BlockType.RESOURCE]:
+                continue
+            source_module = vertex.breadcrumbs.get(CustomAttributes.SOURCE_MODULE)
+            if not source_module:
+                address = self.get_current_address(vertex)
+                vertex.attributes[CustomAttributes.TF_RESOURCE_ADDRESS] = address
+                continue
+            address_prefix = ''
+            for module in source_module:
+                address_prefix += f"{module.get('type')}.{module.get('name')}."
+
+            address = self.get_current_address(vertex, address_prefix)
+            vertex.attributes[CustomAttributes.TF_RESOURCE_ADDRESS] = address
 
 
 def to_list(data):